better controls (wasd kl iop)

+++ srcs/math.c
@@ -6,13 +6,13 @@
 /*   By: axbrisse <axbrisse@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/03 04:41:05 by axbrisse          #+#    #+#             */
-/*   Updated: 2023/02/03 04:41:15 by axbrisse         ###   ########.fr       */
+/*   Updated: 2023/02/05 01:43:06 by axbrisse         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "fdf.h"
 
-int get_min(int n1, int n2)
+double	get_min(double n1, double n2)
 {
 	if (n1 < n2)
 		return (n1);
